Phrase pages, transitions and local label pool
The respective activities for each button in the menu was created and connected, with the exception of the add phrase which was already done. Slide transitions were added for each onclick event. A pool of button labels were created in the strings res file and everytime the menu is loaded the labels for each button are randomly selected.
Note: This pool may be moved to the db if deemed necessary.