Added functionality to add phrase
In the add phrase fragment a connection was made to the firebase db and the phrase entered by the user is added to the respective db child to be moderated (note: at the moment it is added to the child that would be for live use and not just moderation).

To cater for the reading of a random child in the other activities, a count is maintained of the db within the db itself. This count is pulled and incremented before being used as the id for the new phrase being added. The count in the db is then updated.