Added functionality to salt, unsalt and surprise activities
The 'salt' page now randomly pulls a random unfriendly statement from the salt section of the db by using the respective child counter as a limit for the random int

Likewise for the 'sugar' activity.

The 'surprise' activity chooses between the two types randomly and then chooses a random statement in a similar manner as the other activities.

A custom font was added to the text view of the phrases. Size of the text and width of the textView widget were also adjusted.